# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Library API
  version: 1.0.0
  description: >
    Enterprise-grade AI-driven document management system API. Supports document uploads,
    BibTeX metadata auto-extraction, vector-powered semantic search, dual-model AI summarization
    (Llama and Mistral via Spring AI & Ollama), and document RAG conversational chat.
  contact:
    name: Personal Library Architecture Team
    email: support@personallibrary.io
servers:
  - url: /api/v1
    description: Current Environment API Server
  - url: https://api.personallibrary.enterprise/v1
    description: Production Enterprise Gateway
paths:
...
  /documents/{guid}:
    get:
      tags:
        - Documents
      summary: Retrieve document details (Object Page Floorplan)
      description: Returns full metadata, BibTeX formatted representation, dual summaries, and version lineage.
      parameters:
        - name: guid
          in: path
          required: true
          description: Unique identifier of the document
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
        '404':
          description: Document not found

    put:
      tags:
        - Documents
      summary: Overwrite document content and/or metadata in-place (preserves GUID)
      description: >
        Replaces the stored file asset and/or BibTeX metadata while keeping the existing GUID stable.
        Archives the prior state into `versionHistory` and increments `versionNumber`.
      parameters:
        - name: guid
          in: path
          required: true
          description: GUID of the document being overwritten
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Optional replacement physical file. When omitted, only metadata is updated.
                bibtex:
                  type: string
                  description: Optional JSON-encoded replacement BibTeXMetadata string.
      responses:
        '200':
          description: Document overwritten in-place; new version archived to versionHistory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
        '404':
          description: Document not found

    delete:
      tags:
        - Documents
      summary: Delete document record, file asset, and Qdrant vector embeddings
      parameters:
        - name: guid
          in: path
          required: true
          description: GUID of document to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Document and vector embeddings removed
...
  /documents/{guid}/versions/{version}/download:
    get:
      tags:
        - Versioning
        - Documents
      summary: Download historical asset file for a specific version
      description: Streams the original physical file asset preserved at the requested version.
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: path
          required: true
          description: Target version sequence number
          schema:
            type: integer
      responses:
        '200':
          description: Historical document binary attachment
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="document_v1.pdf"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Historical version or document not found
  /documents/{guid}/rollback/{version}:
    post:
      tags:
        - Versioning
        - Documents
      summary: Roll back document state to a designated historical version
      description: >
        Restores file asset, metadata, and dual-model AI summaries from the designated version snapshot,
        archives the active state into the version history, and advances the version sequence counter.
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: path
          required: true
          description: Version sequence number to restore
          schema:
            type: integer
      responses:
        '200':
          description: Document successfully rolled back and updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
        '404':
          description: Document or target version snapshot not found
...
  /documents/{guid}/chat:
    post:
      tags:
        - AI Services
      summary: Submit a conversational query to the document RAG assistant
      description: Retrieves semantic context from Qdrant vector store and generates answer using the Llama model.
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  example: What are the primary experimental results reported?
                chatHistory:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      text:
                        type: string
      responses:
        '200':
          description: RAG response with model metadata and document citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /chat:
    post:
      tags:
        - AI Services
      summary: Document RAG assistant query (convenience endpoint)
      description: Context-aware conversational RAG with citation excerpts. Accepts documentGuid or guid in the JSON body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                documentGuid:
                  type: string
                  format: uuid
                  example: a1b2c3d4-e5f6-4a1b-8c2d-3e4f5a6b7c8d
                guid:
                  type: string
                  format: uuid
                question:
                  type: string
                  example: What are the primary experimental results reported?
                chatHistory:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      text:
                        type: string
      responses:
        '200':
          description: Grounded RAG answer with vector citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
...
components:
  securitySchemes:
    KeycloakBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak OIDC Bearer Token
  schemas:
...
    BibTeXMetadata:
      type: object
      required:
        - entryType
        - bibKey
        - title
        - author
        - year
      properties:
        entryType:
          type: string
          enum: [article, book, inproceedings, techreport, phdthesis, misc]
          example: article
        bibKey:
          type: string
          example: vaswani2017attention
        title:
          type: string
          example: Attention Is All You Need
        author:
          type: string
          example: Vaswani, Ashish and Shazeer, Noam and Parmar, Niki
        year:
          type: string
          example: "2017"
        month:
          type: string
          example: "December"
        journal:
          type: string
          example: Advances in Neural Information Processing Systems
        booktitle:
          type: string
        volume:
          type: string
          example: "30"
        number:
          type: string
        pages:
          type: string
          example: "5998--6008"
        publisher:
          type: string
        edition:
          type: string
        institution:
          type: string
        school:
          type: string
        doi:
          type: string
          example: 10.48550/arXiv.1706.03762
        url:
          type: string
          example: https://arxiv.org/abs/1706.03762
        abstract:
          type: string
        keywords:
          type: string
          example: Attention, Transformer, Deep Learning, NLP
    SummaryRecord:
      type: object
      properties:
        modelName:
          type: string
          example: Ollama Llama 3.3 (70B)
        modelKey:
          type: string
          enum: [llama, mistral]
        summaryText:
          type: string
        createdAt:
          type: string
          format: date-time
        timestamp:
          type: string
          format: date-time
          example: "2026-09-30T05:35:48.000Z"
        durationSeconds:
          type: number
          example: 4.2
        durationFormatted:
          type: string
          example: 0 min 4.2 sec
    DocumentDetail:
      type: object
      properties:
        guid:
          type: string
          format: uuid
        previousVersionGuid:
          type: string
          nullable: true
        versionNumber:
          type: integer
          example: 1
        fileName:
          type: string
          example: attention_is_all_you_need.pdf
        fileSize:
          type: integer
          description: Size in bytes
          example: 2215264
        fileSizeFormatted:
          type: string
          example: 2.1 MB
        format:
          type: string
          enum: [md, docx, pdf, txt, doc, xls, xlsx, ppt, pptx]
          example: pdf
        uploadDate:
          type: string
          format: date-time
        editDate:
          type: string
          format: date-time
        bibtex:
          $ref: '#/components/schemas/BibTeXMetadata'
        bibtexRaw:
          type: string
          description: Pre-formatted LaTeX/BibTeX block string
        summaries:
          type: object
          properties:
            llama:
              $ref: '#/components/schemas/SummaryRecord'
            mistral:
              $ref: '#/components/schemas/SummaryRecord'
        contentExcerpt:
          type: string
        downloadUrl:
          type: string
        versionHistory:
          type: array
          description: Historical immutable revision snapshots archived prior to overwrites or rollbacks
          items:
            $ref: '#/components/schemas/DocumentVersionSnapshot'
    DocumentVersionSnapshot:
      type: object
      properties:
        snapshotGuid:
          type: string
          example: snapshot-b2c3d4e5-v1-archive
        versionNumber:
          type: integer
          example: 1
        fileName:
          type: string
          example: document_v1.pdf
        fileSize:
          type: integer
          example: 524288
        fileSizeFormatted:
          type: string
          example: 512.0 KB
        format:
          type: string
          enum: [md, docx, pdf, txt, doc, xls, xlsx, ppt, pptx]
          example: pdf
        savedAt:
          type: string
          format: date-time
        bibtex:
          $ref: '#/components/schemas/BibTeXMetadata'
        bibtexRaw:
          type: string
        summaries:
          type: object
          properties:
            llama:
              $ref: '#/components/schemas/SummaryRecord'
            mistral:
              $ref: '#/components/schemas/SummaryRecord'
        contentExcerpt:
          type: string
        fullContent:
          type: string
        chunksCount:
          type: integer
          example: 2
        note:
          type: string
          example: Archived prior to upgrade to version 2
...
    ChatResponse:
      type: object
      properties:
        answer:
          type: string
        modelUsed:
          type: string
          example: Llama 3.3 (70B) via Spring AI & Qdrant RAG
        citations:
          type: array
          items:
            type: object
            properties:
              chunkIndex:
                type: integer
              score:
                type: number
              snippet:
                type: string